Place schedule notify calls in telescope code
In relatively recent versions of neovim, notify() must be scheduled in
some circumstances (reading between the lines, I suspect it's to improve
UI responsiveness/reduce stalls or lag). As a result, the telescope
integration was throwing off errors, as reported in issue #4.

This changeset fixes the problem by scheduling those calls to notify, as
inspired by the fix referenced in the changesets here:

https://github.com/neovim/neovim/issues/21052